+++ packages/bridge-controller/src/types.ts
@@ -316,6 +316,7 @@ export type QuoteResponse<
 
 export type BatchSellTradesRequest = {
   quotes: QuoteResponse[];
+  stxEnabled: boolean;
 };
 
 /**