+++ packages/network-controller/src/create-auto-managed-network-client.ts
@@ -100,7 +100,7 @@ export function createAutoManagedNetworkClient<
 }: {
   networkClientId: NetworkClientId;
   networkClientConfiguration: Configuration;
-  getRpcServiceOptions: (
+  getRpcServiceOptions?: (
     rpcEndpointUrl: string,
   ) => RpcServiceOptionsWithDefaults;
   getBlockTrackerOptions?: (