+++ packages/transaction-controller/src/utils/prepare.ts
@@ -33,13 +33,19 @@ export function prepareTransaction(
 }
 
 /**
- * Serializes a transaction object into a hex string.
+ * Serializes transaction data into a hex string.
  *
- * @param transaction - The transaction object.
+ * @param chainId - Chain ID of the transaction.
+ * @param txData - The signed transaction data.
  * @returns The prefixed hex string.
  */
-export function serializeTransaction(transaction: TypedTransaction): Hex {
-  return bytesToHex(transaction.serialize());
+export function serializeTransaction(chainId: Hex, txData: TypedTxData): Hex {
+  return bytesToHex(
+    TransactionFactory.fromTxData(txData, {
+      freeze: false,
+      common: getCommonConfiguration(chainId),
+    }).serialize(),
+  );
 }
 
 /**